Move tracing types to the eth-rpc
