+++ substrate/frame/revive/rpc/Cargo.toml
@@ -26,6 +26,7 @@ codec = { workspace = true, features = ["derive"] }
 derive_more = { workspace = true }
 futures = { workspace = true, features = ["thread-pool"] }
 hex = { workspace = true }
+humantime-serde = { workspace = true }
 jsonrpsee = { workspace = true, features = ["full"] }
 libsqlite3-sys = { workspace = true }
 log = { workspace = true }
@@ -61,8 +62,13 @@ tokio-stream = { workspace = true }
 [dev-dependencies]
 alloy-network = { workspace = true, default-features = true }
 alloy-primitives = { workspace = true, default-features = true }
-alloy-provider = { workspace = true, default-features = true, features = ["debug-api", "reqwest"] }
-alloy-rpc-types = { workspace = true, default-features = true, features = ["trace"] }
+alloy-provider = { workspace = true, default-features = true, features = [
+	"debug-api",
+	"reqwest",
+] }
+alloy-rpc-types = { workspace = true, default-features = true, features = [
+	"trace",
+] }
 alloy-signer-local = { workspace = true, default-features = true }
 env_logger = { workspace = true }
 frame-system = { workspace = true, default-features = true }
